requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
motor==2.5.1
mappymatch==0.4.1
numpy==1.26.4
pydantic[email]
passlib[bcrypt]
pyjwt
pymongo 
